Revert "minor fixes, restructure Tenancy files"
