+++ src/components/portfolio/TenancyItem.tsx
@@ -1,11 +1,12 @@
 import React, { useContext, useEffect, useState, useRef } from 'react'
-import { useNavigate } from 'react-router-dom'
+import { Link, useParams, useNavigate } from 'react-router-dom'
 import { TenancyContext, DELETE_TENANCY } from 'contexts/TenancyContext'
-import { Tenancy as TenancyType, DB_ID } from 'types/global'
+import { Tenancy, DB_ID } from 'types/global'
 import Container from 'react-bootstrap/Container'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import Button from 'react-bootstrap/Button'
+import Breadcrumb from 'react-bootstrap/Breadcrumb'
 import Spinner from 'react-bootstrap/Spinner'
 import { XLg } from 'react-bootstrap-icons'
 import { useTranslation } from 'react-i18next'
@@ -18,18 +19,22 @@ import styled from '@emotion/styled'
 import { deleteTenancyFromStorage } from 'lib/storage'
 import { AuthContext } from 'contexts/AuthContext'
 
+const MyCrumb = styled(Breadcrumb)`
+  ol {
+    background-color: transparent;
+  }
+`
 const ItemContainer = styled(Container)`
   img {
     max-width: 85vw;
   }
   margin-bottom: 40px;
 `
-
-type TenancyDetailsProps = {
-  item: TenancyType
+type TheItemProps = {
+  item: Tenancy
 }
 
-const TenancyDetails = ({ item }: TenancyDetailsProps) => {
+const TheItem = ({ item }: TheItemProps) => {
   const url =
     process.env.NODE_ENV === 'production'
       ? `https://maps.googleapis.com/maps/api/streetview?size=400x400&location=${item.y},${item.x}&fov=70&pitch=0&key=${process.env.REACT_APP_GOOGLE_API_KEY}`
@@ -122,4 +127,45 @@ const TenancyDetails = ({ item }: TenancyDetailsProps) => {
   )
 }
 
-export default TenancyDetails
+const TenancyItem = () => {
+  const { id } = useParams()
+  const { t } = useTranslation()
+  const { tenancies } = useContext(TenancyContext)
+
+  const [item, setItem] = useState<Tenancy | null>(null)
+  const [error, setError] = useState('')
+
+  useEffect(() => {
+    const item = tenancies.filter((obj) => obj.id === id)[0]
+    if (item) {
+      setError('')
+      setItem(item)
+    } else {
+      setItem(null)
+      setError('Item not found :(')
+    }
+  }, [tenancies, id])
+
+  return (
+    <>
+      <MyCrumb>
+        <Breadcrumb.Item linkAs={Link} linkProps={{ to: '/' }}>
+          {t('tenancy_item_breadcrumb_back_to_list')}
+        </Breadcrumb.Item>
+        <Breadcrumb.Item active>
+          {error && 'ERROR'}
+          {item && item.tekst}
+        </Breadcrumb.Item>
+      </MyCrumb>
+      {error && (
+        <Container>
+          <Row>
+            <Col>{error}</Col>
+          </Row>
+        </Container>
+      )}
+      {item && <TheItem item={item} />}
+    </>
+  )
+}
+export default TenancyItem
